requirements.txt
Flask==2.0.3
mediapipe==0.8.9.1
numpy==1.21.6
opencv-python==4.5.5.64
pandas==1.3.5
#Additional
pymongo==4.3.3
dnspython==2.3.0
pyrebase4==4.6.0